requirements.txt
asgiref==3.2.7
asn1crypto==0.24.0
beautifulsoup4==4.9.0
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
config==0.4.0
cryptography==3.2
defusedxml==0.6.0
Django==2.2.20
django-allauth==0.41.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
enum34==1.1.6
idna==2.6
ipaddress==1.0.19
jsonfield==2.0.2
oauthlib==3.1.0
pdfkit-async==0.7.1
Pillow==8.1.1
psycopg2
pycparser==2.18
pyOpenSSL==17.5.0
python3-openid==3.1.0
pytz==2018.3
requests==2.20.0
requests-oauthlib==1.3.0
six==1.11.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.24.2
